+++ public/admin/includes/header.php
@@ -14,8 +14,7 @@ content="width=device-width, initial-scale=1.0">
 
 <title><?php echo SITE_NAME; ?></title>
 
-<link rel="icon"
-href="/assets/images/logo/logo.png">
+<link rel="icon" href="<?= ASSET_URL ?>/images/logo/logo.png">
 
 <link
 href="https://cdn.jsdelivr.net/npm/bootstrap@5.3.7/dist/css/bootstrap.min.css"
@@ -29,7 +28,7 @@ href="https://cdnjs.cloudflare.com/ajax/libs/font-awesome/6.7.2/css/all.min.css"
 href="https://fonts.googleapis.com/css2?family=Poppins:wght@300;400;500;600;700&display=swap"
 rel="stylesheet">
 
-<link rel="stylesheet" href="/admin/assets/css/admin.css">
+<link rel="stylesheet" href="<?= ASSET_URL ?>/css/admin.css">
 
 </head>